requirements.txt
pandas==1.5.3
pyperclip==1.8.2
streamlit==1.21.0
streamlit-option-menu==0.3.2
mysql-connector-python==8.0.33
